Address command palette review follow-ups
